shift octave fatto. ora abbiamo A4 = 440

+++ Rosalba break synth/Source/Data/OscData.h
@@ -26,5 +26,5 @@ public:
     void setOctave(const int choice);
 
 private:
-    float shiftOctave;
+    double shiftOctave;
 };
